refactor(braille): split braille.input into constants, gesture and inputHandler submodules
