+++ source/braille/input/inputHandler.py
@@ -1,7 +1,9 @@
 # A part of NonVisual Desktop Access (NVDA)
-# This file is covered by the GNU General Public License.
-# See the file COPYING for more details.
-# Copyright (C) 2012-2025 NV Access Limited, Rui Batista, Babbage B.V., Julien Cochuyt, Leonard de Ruijter
+# Copyright (C) 2012-2026 NV Access Limited, Rui Batista, Babbage B.V., Julien Cochuyt, Leonard de Ruijter
+# This file may be used under the terms of the GNU General Public License, version 2 or later, as modified by the NVDA license.
+# For full terms and any additional permissions, see the NVDA license file: https://github.com/nvaccess/nvda/blob/master/copying.txt
+
+"""Handler for braille input."""
 
 import time
 from typing import Optional, List, Set
@@ -22,26 +24,13 @@ from baseObject import AutoPropertyObject
 import keyLabels
 import struct
 
-
-"""Framework for handling braille input from the user.
-All braille input is represented by a {BrailleInputGesture}.
-Normally, all that is required is to create and execute a L{BrailleInputGesture},
-as there are built-in gesture bindings for braille input.
-"""
-
-FALLBACK_TABLE = config.conf.getConfigValidation(("braille", "inputTable")).default
-"""Table to use if the input table configuration is invalid."""
-
-DOT7 = 1 << 6
-DOT8 = 1 << 7
-#: This bit flag must be added to all braille cells when using liblouis with dotsIO.
-LOUIS_DOTS_IO_START = 0x8000
-#: The start of the Unicode braille range.
-#: @type: int
-UNICODE_BRAILLE_START = 0x2800
-#: The Unicode braille character to use when masking cells in protected fields.
-#: @type: str
-UNICODE_BRAILLE_PROTECTED = "⣿"  # All dots down
+from .constants import (
+	FALLBACK_TABLE,
+	LOUIS_DOTS_IO_START,
+	UNICODE_BRAILLE_START,
+	UNICODE_BRAILLE_PROTECTED,
+)
+from .gesture import formatDotNumbers
 
 
 class BrailleInputHandler(AutoPropertyObject):
@@ -59,30 +48,34 @@ class BrailleInputHandler(AutoPropertyObject):
 	def __init__(self):
 		super().__init__()
 		self._table: brailleTables.BrailleTable = brailleTables.getTable(FALLBACK_TABLE)
-		#: A buffer of entered braille cells so that state set by previous cells can be maintained;
-		#: e.g. capital and number signs.
 		self.bufferBraille = []
-		#: The text translated so far from the cells in L{bufferBraille}.
+		"""A buffer of entered braille cells so that state set by previous cells can be maintained;
+		e.g. capital and number signs.
+		"""
 		self.bufferText = ""
-		#: Indexes of cells which produced text.
-		#: For example, this includes letters and numbers, but not number signs,
-		#: since a number sign by itself doesn't produce text.
-		#: This is used when erasing cells to determine when to backspace an actual character.
+		"""The text translated so far from the cells in :attr:`bufferBraille`."""
 		self.cellsWithText = set()
-		#: The cells in L{bufferBraille} that have not yet been translated
-		#: or were translated but did not produce any text.
-		#: This is used to show these cells to the user while they're entering braille.
-		#: This is a string of Unicode braille.
+		"""Indexes of cells which produced text.
+		For example, this includes letters and numbers, but not number signs,
+		since a number sign by itself doesn't produce text.
+		This is used when erasing cells to determine when to backspace an actual character.
+		"""
 		self.untranslatedBraille = ""
-		#: The position in L{brailleBuffer} where untranslated braille begins.
+		"""The cells in :attr:`bufferBraille` that have not yet been translated
+		or were translated but did not produce any text.
+		This is used to show these cells to the user while they're entering braille.
+		This is a string of Unicode braille.
+		"""
 		self.untranslatedStart = 0
-		#: The user's cursor position within the untranslated braille.
-		#: This enables the user to move within the untranslated braille.
+		"""The position in :attr:`bufferBraille` where untranslated braille begins."""
 		self.untranslatedCursorPos = 0
-		#: The time at which uncontracted characters were sent to the system.
+		"""The user's cursor position within the untranslated braille.
+		This enables the user to move within the untranslated braille.
+		"""
 		self._uncontSentTime = None
-		#: The modifiers currently being held virtually to be part of the next braille input gesture.
+		"""The time at which uncontracted characters were sent to the system."""
 		self.currentModifiers = set()
+		"""The modifiers currently being held virtually to be part of the next braille input gesture."""
 		self.handlePostConfigProfileSwitch()
 		config.post_configProfileSwitch.register(self.handlePostConfigProfileSwitch)
 
@@ -97,7 +90,7 @@ class BrailleInputHandler(AutoPropertyObject):
 		self._table = table
 		config.conf["braille"]["inputTable"] = table.fileName
 
-	# Provided by auto property: L{_get_currentFocusIsTextObj}
+	# Provided by auto property: :meth:`_get_currentFocusIsTextObj`
 	currentFocusIsTextObj: bool
 
 	def _get_currentFocusIsTextObj(self):
@@ -106,7 +99,7 @@ class BrailleInputHandler(AutoPropertyObject):
 			not focusObj.treeInterceptor or focusObj.treeInterceptor.passThrough
 		)
 
-	# Provided by auto property: L{_get_useContractedForCurrentFocus}
+	# Provided by auto property: :meth:`_get_useContractedForCurrentFocus`
 	useContractedForCurrentFocus: bool
 
 	def _get_useContractedForCurrentFocus(self):
@@ -115,8 +108,8 @@ class BrailleInputHandler(AutoPropertyObject):
 	def _translate(self, endWord: bool) -> bool:
 		"""Translate buffered braille up to the cursor.
 		Any text produced is sent to the system.
-		@param endWord: C{True} if this is the end of a word, C{False} otherwise.
-		@return: C{True} if translation produced text, C{False} if not.
+		:param endWord: ``True`` if this is the end of a word, ``False`` otherwise.
+		:return: ``True`` if translation produced text, ``False`` if not.
 		"""
 		assert not self.useContractedForCurrentFocus or endWord, (
 			"Must only translate contracted at end of word"
@@ -150,7 +143,7 @@ class BrailleInputHandler(AutoPropertyObject):
 			if self.currentModifiers or not self.currentFocusIsTextObj:
 				if len(newText) > 1:
 					# Emulation of multiple characters at once is unsupported
-					# Clear newText, so this function returns C{False} if not at end of word
+					# Clear newText, so this function returns ``False`` if not at end of word
 					newText = ""
 				else:
 					self.emulateKey(newText)
@@ -173,10 +166,10 @@ class BrailleInputHandler(AutoPropertyObject):
 
 		return False
 
-	def _translateForReportContractedCell(self, pos):
-		"""Translate text for current input as required by L{_reportContractedCell}.
-		@return: The previous translated text.
-		@rtype: str
+	def _translateForReportContractedCell(self, pos) -> str:
+		"""Translate text for current input as required by :meth:`_reportContractedCell`.
+
+		:return: The previous translated text.
 		"""
 		cells = self.bufferBraille[: pos + 1]
 		data = "".join([chr(cell | LOUIS_DOTS_IO_START) for cell in cells])
@@ -189,14 +182,13 @@ class BrailleInputHandler(AutoPropertyObject):
 		self.bufferText = text
 		return oldText
 
-	def _reportContractedCell(self, pos):
+	def _reportContractedCell(self, pos) -> bool:
 		"""Report a guess about the character(s) produced by a cell of contracted braille.
 		It's not possible to report the exact characters because later cells might change text produced by earlier cells.
 		However, it's helpful for the user to have a rough idea.
 		For example, in English contracted braille, "alw" is the contraction for "always".
 		As the user types "alw", the characters a, l, w will be spoken.
-		@return: C{True} if a guess was reported, C{False} if not (e.g. a number sign).
-		@rtype: bool
+		:return: ``True`` if a guess was reported, ``False`` if not (e.g. a number sign).
 		"""
 		oldText = self._translateForReportContractedCell(pos)
 		oldTextLen = len(oldText)
@@ -298,7 +290,7 @@ class BrailleInputHandler(AutoPropertyObject):
 
 	def _updateUntranslated(self):
 		"""Update the untranslated braille to be shown to the user.
-		If the display will not otherwise be updated, L{updatedisplay} should be called after this.
+		If the display will not otherwise be updated, :meth:`updateDisplay` should be called after this.
 		"""
 		if api.isTypingProtected():
 			self.untranslatedBraille = UNICODE_BRAILLE_PROTECTED * (
@@ -381,8 +373,8 @@ class BrailleInputHandler(AutoPropertyObject):
 		"""Emulates a key using the keyboard emulation system.
 		If emulation fails (e.g. because of an unknown key), a debug warning is logged
 		and the system falls back to sending unicode characters.
-		@param withModifiers: Whether this key emulation should include the modifiers that are held virtually.
-			Note that this method does not take care of clearing L{self.currentModifiers}.
+		:param withModifiers: Whether this key emulation should include the modifiers that are held virtually.
+			Note that this method does not take care of clearing :attr:`currentModifiers`.
 		"""
 		if withModifiers:
 			# The emulated key should be the last item in the identifier string.
@@ -402,7 +394,7 @@ class BrailleInputHandler(AutoPropertyObject):
 
 	def sendChars(self, chars: str):
 		"""Sends the provided unicode characters to the system.
-		@param chars: The characters to send to the system.
+		:param chars: The characters to send to the system.
 		"""
 		inputs = []
 		chars = "".join(
@@ -472,124 +464,6 @@ class BrailleInputHandler(AutoPropertyObject):
 				self._table = brailleTables.getTable(FALLBACK_TABLE)
 
 
-#: The singleton BrailleInputHandler instance.
-handler: Optional[BrailleInputHandler] = None
-
-
-def initialize():
-	global handler
-	handler = BrailleInputHandler()
-	log.info("Braille input initialized")
-
-
-def terminate():
-	global handler
-	handler = None
-
-
-def formatDotNumbers(dots: int):
-	out = []
-	for dot in range(8):
-		if dots & (1 << dot):
-			out.append(str(dot + 1))
-	return " ".join(out)
-
-
 def speakDots(dots: int):
 	# Translators: Used when reporting braille dots to the user.
 	speech.speakMessage(_("dot") + " " + formatDotNumbers(dots))
-
-
-class BrailleInputGesture(inputCore.InputGesture):
-	"""Input (dots and/or space bar) from a braille keyboard.
-	This could either be as part of a braille display or a stand-alone unit.
-	L{dots} and L{space} should be set appropriately.
-	"""
-
-	#: Bitmask of pressed dots.
-	#: @type: int
-	dots = 0
-
-	#: Whether the space bar is pressed.
-	#: @type: bool
-	space = False
-
-	shouldPreventSystemIdle = True
-
-	def _makeDotsId(self):
-		items = ["dot%d" % (i + 1) for i in range(8) if self.dots & (1 << i)]
-		if self.space:
-			items.append("space")
-		return "bk:" + "+".join(items)
-
-	#: The generic gesture identifier for space plus any dots.
-	#: This could be used to bind many braille commands to a single script.
-	GENERIC_ID_SPACE_DOTS = inputCore.normalizeGestureIdentifier("bk:space+dots")
-	#: The generic gesture identifier for any dots.
-	#: This is used to bind entry of braille text to a single script.
-	GENERIC_ID_DOTS = inputCore.normalizeGestureIdentifier("bk:dots")
-
-	def _get_identifiers(self):
-		if self.space and self.dots:
-			return (self._makeDotsId(), self.GENERIC_ID_SPACE_DOTS)
-		elif self.dots in (DOT7, DOT8, DOT7 | DOT8):
-			# Allow bindings to dots 7 and/or 8 by themselves.
-			return (self._makeDotsId(), self.GENERIC_ID_DOTS)
-		elif self.dots or self.space:
-			return (self.GENERIC_ID_DOTS,)
-		else:
-			return ()
-
-	@classmethod
-	def _makeDisplayText(cls, dots: int, space: bool):
-		out = ""
-		if space and dots:
-			# Translators: Reported when braille space is pressed with dots in input help mode.
-			out = _("space with dot")
-		elif dots:
-			# Translators: Reported when braille dots are pressed in input help mode.
-			out = _("dot")
-		elif space:
-			# Translators: Reported when braille space is pressed in input help mode.
-			out = _("space")
-		if dots:
-			out += " " + formatDotNumbers(dots)
-		return out
-
-	def _get_displayName(self):
-		if not self.dots and not self.space:
-			return None
-		return self._makeDisplayText(self.dots, self.space)
-
-	@classmethod
-	def getDisplayTextForIdentifier(cls, identifier: str):
-		assert isinstance(identifier, str)
-		# Translators: Used when describing keys on a braille keyboard.
-		source = _("braille keyboard")
-		if identifier == cls.GENERIC_ID_SPACE_DOTS:
-			# Translators: Used to describe the press of space
-			# along with any dots on a braille keyboard.
-			return (source, _("space with any dots"))
-		if identifier == cls.GENERIC_ID_DOTS:
-			# Translators: Used to describe the press of any dots
-			# on a braille keyboard.
-			return (source, _("any dots"))
-		# Example identifier: bk:space+dot1+dot2
-		# Strip the bk: prefix.
-		partsStr = identifier.split(":", 1)[1]
-		parts = partsStr.split("+")
-		dots = 0
-		space = False
-		for part in parts:
-			if part == "space":
-				space = True
-			else:
-				# Example part: "dot1"
-				# Get the dot number and make it 0 based instead of 1 based.
-				dot = int(part[3]) - 1
-				# Update the dots bitmask.
-				dots += 1 << dot
-		return (source, cls._makeDisplayText(dots, space))
-
-
-inputCore.registerGestureSource("bk", BrailleInputGesture)
